added products to category 3 and images

+++ src/main/java/org/teamraccoon/dreamfusion/products/Product.java
@@ -44,7 +44,7 @@ public class Product {
     private String productImage;
 
     @Column
-    private Long price;
+    private float price;
 
     @Column
     @ManyToMany(fetch = FetchType.EAGER)
